+++ flask_unchained/__init__.py
@@ -8,7 +8,7 @@
     :license: MIT, see LICENSE for more details
 """
 
-__version__ = '0.3.2'
+__version__ = '0.4.0'
 
 
 from .app_factory import AppFactory